150686 by jwiele on 2021/06/25 (pair: jshimkus) VDOUP-6 Remove PRI* print macros.
	The PRI* print macros can be replaced with standard format
	directives with appropriate casts for platform defined integer
	types. For example, uint64_t may be defined as "unsigned long"
	or "unsigned long long" but the latter is guaranteed to always
	be at least 64 bits, so a uint64_t quantity may be printed
	safely by casting to "unsigned long long and formatting with
	the "%llu" directive. See change 150333 for more details.

	.../blockMapFormat.c
	.../slabDepotFormat.c
	.../vdoComponent.c
	.../vdoComponentStates.c
	.../fileLayer.c
	.../vdoAudit.c
	.../vdoDebugMetadata.c
	.../vdoDumpBlockMap.c
	.../vdoDumpConfig.c
	.../vdoDumpMetadata.c
	.../vdoFormat.c
	.../vdoRegenerateGeometry.c
	Removed use of PRI* print macros, replacing them with standard
	  print format characters, adding casts where necessary for
	  platform dependent types.

+++ utils/vdo/fileLayer.c
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301, USA. 
  *
- * $Id: //eng/linux-vdo/src/c++/vdo/user/fileLayer.c#17 $
+ * $Id: //eng/linux-vdo/src/c++/vdo/user/fileLayer.c#18 $
  */
 
 #include "fileLayer.h"
@@ -163,8 +163,8 @@ static int fileReader(PhysicalLayer           *header,
     return VDO_OUT_OF_RANGE;
   }
 
-  uds_log_debug("FL: Reading %zu blocks from block %" PRIu64,
-		blockCount, startBlock);
+  uds_log_debug("FL: Reading %zu blocks from block %llu",
+		blockCount, (unsigned long long) startBlock);
 
   // Make sure we cast so we get a proper 64 bit value on the calculation
   char *alignedBuffer;
@@ -197,8 +197,8 @@ static int fileWriter(PhysicalLayer           *header,
     return VDO_OUT_OF_RANGE;
   }
 
-  uds_log_debug("FL: Writing %zu blocks from block %" PRIu64,
-		blockCount, startBlock);
+  uds_log_debug("FL: Writing %zu blocks from block %llu",
+		blockCount, (unsigned long long) startBlock);
 
   // Make sure we cast so we get a proper 64 bit value on the calculation
   size_t bytes = blockCount * VDO_BLOCK_SIZE;
