+++ utils/vdo/vdodumpblockmap.c
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301, USA. 
  *
- * $Id: //eng/linux-vdo/src/c++/vdo/user/vdoDumpBlockMap.c#20 $
+ * $Id: //eng/linux-vdo/src/c++/vdo/user/vdoDumpBlockMap.c#21 $
  */
 
 #include <err.h>
@@ -121,23 +121,23 @@ static int dumpLBN(void)
   enum block_mapping_state state;
   int result = findLBNMapping(vdo, lbn, &pbn, &state);
   if (result != VDO_SUCCESS) {
-    warnx("Could not read mapping for lbn %" PRIu64, lbn);
+    warnx("Could not read mapping for lbn %llu", (unsigned long long) lbn);
     return result;
   }
 
-  printf("%" PRIu64 "\t", lbn);
+  printf("%llu\t", (unsigned long long) lbn);
   switch (state) {
   case VDO_MAPPING_STATE_UNMAPPED:
-    printf("unmapped   \t%" PRIu64 "\n", pbn);
+    printf("unmapped   \t%llu\n", (unsigned long long) pbn);
     break;
 
   case VDO_MAPPING_STATE_UNCOMPRESSED:
-    printf("mapped     \t%" PRIu64 "\n", pbn);
+    printf("mapped     \t%llu\n", (unsigned long long) pbn);
     break;
 
   default:
-    printf("compressed \t%" PRIu64 " slot %u\n",
-           pbn, vdo_get_slot_from_state(state));
+    printf("compressed \t%llu slot %u\n",
+           (unsigned long long) pbn, vdo_get_slot_from_state(state));
     break;
   }
 
@@ -155,9 +155,10 @@ static int dumpBlockMapEntry(struct block_map_slot    slot,
                              enum block_mapping_state state)
 {
   if ((state != VDO_MAPPING_STATE_UNMAPPED) || (pbn != VDO_ZERO_BLOCK)) {
-    printf("PBN %" PRIu64 "\t slot %u\t height %u\t"
-           "-> PBN %" PRIu64 " (compression state %u)\n",
-           slot.pbn, slot.slot, height, pbn, state);
+    printf("PBN %llu\t slot %u\t height %u\t"
+           "-> PBN %llu (compression state %u)\n",
+           (unsigned long long) slot.pbn, slot.slot, height,
+           (unsigned long long) pbn, state);
   }
   return VDO_SUCCESS;
 }
